WIP: first implementation of the idea
Here I plumb in a path to call wombat to get the convex regions
and give that off the mpi distributed.

Hard to tell for it is working because there are very noticable
region bound artifacts with mpi distributed volume that preexist
this demonstration, and with AMR there are many more regions.