finish the small simple example
seems to be more or less correct. An observation is that
remote value treatment is tricky, NAN is the worst, the original
approach of setting remotes to 0 and letting the LUT handle it
is pretty good but boundaries are fairly visible where they
transition to the 0 color. Setting remote values to local value is
quite close, but there are still visible boundaries. Setting it to
the actual value is closer still but even that has some artifacts.